Originally this was supposed to be a very minor update. It appears that I forgot to commit the last update. Now up to revision 140 of the original project. ColorSpace support for YCCK. Better determination on if a built in font is italic or not. Added ability to get names from a font's NameTable (this functionality exists in TrueTypeFont, and TTFFont as well). Head table adjustment for TrueTypeFont. Added the ability to load a TTF font from a file. Added helper type for a read-only Vector. Added additional check in PDFRenderer so when a masked image is retrieved, and the fill state is retrieved, it doesn't crash because there is no solid color. (Not actually implemented) Added ability for GraphicsImpl (for GraphicsINternalGraphics) to have properties (though none exist right now).

+++ PDFRenderer/src/com/sun/pdfview/PDFDocCharsetEncoder.java
@@ -72,14 +72,14 @@ public class PDFDocCharsetEncoder extends CharsetEncoder
     static
     {
     	int len = PDFStringUtil.PDF_DOC_ENCODING_MAP.length;
-        for (byte i = 0; i < len; ++i)
+        for (int i = 0; i < len; ++i)
         {
             final short c = (short)PDFStringUtil.PDF_DOC_ENCODING_MAP[i];
             final boolean identical = (c == i);
             IDENT_PDF_DOC_ENCODING_MAP[i] = identical;
             if (!identical)
             {
-                EXTENDED_TO_PDF_DOC_ENCODING_MAP.put(new Short(c), new Byte(i));
+                EXTENDED_TO_PDF_DOC_ENCODING_MAP.put(new Short(c), new Byte((byte)i));
             }
         }
     }
